+++ angular_compiler/lib/src/analyzer.dart
@@ -1,5 +1,4 @@
 export 'analyzer/common.dart';
-export 'analyzer/di/component.dart';
 export 'analyzer/di/dependencies.dart';
 export 'analyzer/di/injector.dart';
 export 'analyzer/di/modules.dart';
@@ -8,3 +7,4 @@ export 'analyzer/di/tokens.dart';
 export 'analyzer/link.dart';
 export 'analyzer/reflector.dart';
 export 'analyzer/types.dart';
+export 'analyzer/view/directive.dart';